Allow the planet to move/rotate
Now, the planet can rotate, move translationally, and scale up or down in size along the X and Y axes. Also, I have added the JOML library as a dependency, allowed for matrix operations, expanded functionality to listen to keyboard commands, and changed the shape to the color green instead of yellow.